src/examples/python/openai_sdk_0.28.x/requirements.txt
openai>=0.28.1, <0.29
python_dotenv>=1.0.0, <2.0.0
langchain
black
ruff